requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.6
django-appconf==1.1.0
django-autoslug==1.9.9
django-csp==4.0
django-environ==0.12.0
django-imagekit==5.0.0
packaging==25.0
pilkit==3.0
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
whitenoise==6.11.0
